svn_rev_183
restructure ppc startup process. Instead of doing the whole
initialization sequence in flash, we now initialize sdram very
early in the start sequence and relocate u-boot into RAM.
Also U-Boot is now linked to address 0x0 and is also executed
from there. This gives the advantage that we see the correct
offsets in the objdump and we do not need to hassle with
relocation anymore.
The code is still position independent and compiled with -fPIC,
but this could be removed.