loadb: create file
Allow loadb and loady to create file if file not
present. This will allow for downloading a file
to filesystem and cp or doing other operations on
the same. Making this as an option instead of a
default behavior ensures that users intend to
create file when they use -c option

Signed-off-by: Nishanth Menon <x0nishan@ti.com>