Ethernet: write MAC address to hardware when it's changed
Instead of writing the MAC address into the hardware on device open time
write it to the hardware when the user changes the parameter. This
way a user can change the MAC address in the hardware without actually
using the device. This helps Linux Network drivers which expect a valid
MAC address on startup.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>