pcm038: initialize PLLs in one step
We used to initialize the PLLs on PCM038 in two
steps. The first was to initialize a basic setup
so that all peripheral clocks run at their final
speed and the second step was to push the processor
to full speed after the PMIC has been initialized
for the higher voltage.
Do this in one step instead to get an easier setup.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>