nand bb: Fix writing of nand page unaligned data
We used to write the data in nand_bb_write directly to the
NAND flash. If we do not write a whole NAND page at once, this
resulted in multiple writes of one page which corrupted the
ecc data.
Fix this by collecting 4096 bytes of data before actually
writing the data to the flash.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>