A9M2440: Generic code should support generic hardware
The generic A9M2440 comes with a CS8900 network controller at chip select 5.
There are custom specific variants available without this device, but this
case will be handled in a separate baseboard file.

The current Linux-2.6.31 CS8900 network driver is a very poor piece of
code. It needs some patches to also work with this platform. So, I add
two kernel parameters to be able to use this network device also in the NFS
rootfs case (to setup media and MAC).

Signed-off-by: Juergen Beisert <juergen@kreuzholzen.de>