A9M2440: Use external GPIO settings to setup the SDRAM
The generic A9M2440 comes in many SDRAM variantes. Two GPIO pins describing
the populated SDRAM devices. This patch adds some code to query the setting
and setup a correct SDRAM configuration out of four possible variants.

Due to the lack of hardware the 16 MiB and 128 MiB case is untested yet.

Due to the 32 MiB configuration we must link the u-boot-v2 always to
address 0x31fc0000. Even for the 64 MiB and 128 MiB configuration.

Signed-off-by: Juergen Beisert <juergen@kreuzholzen.de>