Move private settings and configuration into the C source
Include the board config first to be able to configure the memory management
in the documented way. If not used, the defaults are used.
On the other hand, there is no need to pollute the other source files with
these local management settings. So, move them from the header into the
C source file.

Signed-off-by: Juergen Beisert <jbe@pengutronix.de>