imx35pdk: Detect and set silicon and board revision for boot params.
The kernel for the imx35 pdk doesn't do any detection for board and chip revision. This patch passes the info to the kernel so that v2 boards/chips boot properly.

imx35pdk: make COMPARE pad a GPIO1_5 output

This pad is an interrupt source, active low for one of the external ICs.

fix up/remove unnecessary white space.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>