mx25 3ds: fix fec initialization
The fec network interface initialization depends on the initialization of the
PMIC. Once the MC34704 driver is registered we can enable the PHY power supply,
and go on with the PHY initialization.

While we are at it convert the hard-coded GPIO registers access to the general
GPIO API for shorter and clearer code.

Signed-off-by: Baruch Siach <baruch@tkos.co.il>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>