Avoid warnings by using format(__printf__)
When building sandbox, barebox is built with -Dprintf=barebox_printf as
to not collide with the printf provided by libc.  This would also match
the format(printf) function __attribute__.

Since gcc documents that __printf__ can be used instead of printf as a
format attribute, use this instead and avoid a lot of noisy warnings.

NB: this relates to 6b082cfe9f9b5b2bea294918ad916c739490cea7 which was
an earlier attempt at solving this, which got reverted due to other
regressions.

Signed-off-by: Loïc Minier <loic.minier@linaro.org>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>