drivers/video: remove pxafb enable on load
As pxafb can rely on a PWM to control backlight, and because
driver dependencies are hard to deal with, remove automatic
enable of PXAFB on probe.

The user should in its environment do a :
 - fb0.enable=1

This way, the PWM has been probed and is ready to work, and
the pxafb backlight control works.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>